+++ src/components/ui/button.tsx
@@ -3,7 +3,7 @@ import { ButtonHTMLAttributes } from 'react';
 
 interface IButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   // 추가적인 프롭스가 필요한 경우 여기에 정의
-  color: 'main' | 'discord' | '';
+  color: 'main' | 'discord' | 'gray';
   size: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'wide';
 }
 
@@ -18,7 +18,8 @@ const sizes = {
 
 const colors = {
   main: ' bg-main text-White',
-  discord: 'bg-discord '
+  discord: 'bg-discord ',
+  gray: 'bg-tableBackground border'
 };
 
 const Button: React.FunctionComponent<IButtonProps> = ({
