+++ src/components/ui/input.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import clsx from 'clsx';
 import { InputHTMLAttributes } from 'react';
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
@@ -15,8 +16,10 @@ export default function Input({ name, type, icon, invalid, labelRequired, ...pro
       <div className='pointer-events-none absolute inset-y-0 start-0 flex items-center ps-3'></div>
       <input
         id={name}
-        className='h-[30px] w-full appearance-none rounded-md border px-4 py-2 text-sm font-semibold text-black placeholder-gray-400 shadow-sm focus:border-main focus:outline-none 
-              focus:ring-2 focus:ring-main [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none'
+        className={clsx(
+          `h-[30px] w-full rounded-md border px-4 py-2 text-sm font-semibold text-black placeholder-gray-400 shadow-sm focus:border-main focus:outline-none 
+        focus:ring-2 focus:ring-main`
+        )}
         {...props}
       />
       <div className='absolute bottom-2 end-2 flex items-center gap-x-1 '>{icon}</div>
