--- src/components/ui/radio-group.tsx
@@ -1,7 +1,0 @@
-interface IRadioGroupProps {}
-
-const RadioGroup: React.FunctionComponent<IRadioGroupProps> = ({}) => {
-  return <></>;
-};
-
-export default RadioGroup;
